[generate_opcheck_tests] Always print a repro
On failure of a test, we will always print a "repro". This repro isn't
really runnable but gives the user a sense of how to actually reproduce
the test without the test suite, because using the test suite is a bit
convoluted.

If the user passes PYTORCH_OPCHECK_PRINT_BETTER_REPRO, we will print a
fuller repro that saves the exact problematic test inputs to disk and
reads them back out.

Test Plan:
- expecttests on the generate_repro helper function
- tried this out locally.

ghstack-source-id: 8641fc32e6463bb405dac47eae80cc555cb97d51
Pull Request resolved: https://github.com/pytorch/pytorch/pull/109640