[generate_opcheck_tests] tests should ignore meta/FakeTensors
These tests generally don't work on meta tensors because they need to
compare the data of the Tensors. For example, SchemaCheckMode errors out
if any inputs are meta or Fake because it needs to check their storages
to see if any mutation occurred and those do not have storages.

Test Plan:
- new tests

ghstack-source-id: 0fc2daa8391ee0fa236857d64f229e30c2ef18f0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/109641